pyproject.toml
[tool.poetry]
name = "prac-diplom-filler"
version = "0.1.0"
description = ""
authors = ["Vitaly Chernyshov <81708357+vitaliy-chernyshov@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "prac_diplom_filler"}]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "9.4.0"
tqdm = "^4.65.0"
faker = "^17.6.0"
aiohttp = "^3.8.4"


[tool.poetry.group.dev.dependencies]
flake8-black = "^0.3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
